🧪 test: split streaming tests into LLMCallerStreamingTests
Move the 3 streaming-path tests (`emitsAgentOutputStreamPerDelta…`,
`streamSnapshotsArrayIsMonotonic…`, `streamRetryOnParseFailure…`) out
of `LLMCallerTests` into a new file to stay under the
type-body-length lint budget (250 lines). No behaviour change.

Refs #119.
